+++ demos/manual/src/main/java/is/codion/framework/demos/manual/store/minimal/ui/StoreDemo.java
@@ -5,7 +5,7 @@ package is.codion.framework.demos.manual.store.minimal.ui;
 
 import is.codion.common.db.database.Database;
 import is.codion.common.user.User;
-import is.codion.dbms.h2database.H2DatabaseFactory;
+import is.codion.dbms.h2.H2DatabaseFactory;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.db.local.LocalEntityConnectionProvider;
 import is.codion.framework.demos.manual.store.minimal.domain.Store;